More simplifications and dead code removal.
