+++ DEVLOG.md
@@ -234,6 +234,8 @@ _---J. Ian Lindsay_
     280220    2880   29832  312932   Integrator is starting to not care about sample type.
     279836    2880   20040  302756   The big drop. Note below.
     279564    2880   19880  302324   More removal of redundant members.
+    279500    2880   19264  301644   Removal of value cache in IMU classes.
+    278964    2880   19264  301108   Removal of dead code.
 
 Lots of things are broken. There are a few stubs, and much dead code. IMU pointer
 maps are contained within a const class and array'd out as non-static members.
