+++ Makefile.emu
@@ -59,8 +59,6 @@ LDFLAGS += -L. -L$(OUTPUT_PATH)
 SOURCES_CPP   = src/main-emu.cpp
 SOURCES_CPP  += src/Digitabulum/CPLDDriver/CPLDDriver.cpp
 SOURCES_CPP  += src/Digitabulum/LSM9DS1/LSM9DS1.cpp
-SOURCES_CPP  += src/Digitabulum/LSM9DS1/LSM9DS1_AG.cpp
-SOURCES_CPP  += src/Digitabulum/LSM9DS1/LSM9DS1_M.cpp
 SOURCES_CPP  += src/Digitabulum/ManuLegend/SensorFrame.cpp
 SOURCES_CPP  += src/Digitabulum/ManuLegend/Integrator.cpp
 SOURCES_CPP  += src/Digitabulum/ManuLegend/ManuManager.cpp
